pyproject.toml
[project]
name = "hunyuan3d-2"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "accelerate>=1.3.0",
    "diffusers>=0.32.2",
    "einops>=0.8.0",
    "fastapi>=0.115.6",
    "gradio>=5.12.0",
    "imageio>=2.37.0",
    "llvmlite>=0.44.0",
    "ninja>=1.11.1.3",
    "numba>=0.61.0",
    "numpy>=2.1.3",
    "omegaconf>=2.3.0",
    "onnxruntime>=1.20.1",
    "opencv-python>=4.11.0.86",
    "pybind11>=2.13.6",
    "pyglet<2",
    "pygltflib>=1.16.3",
    "pymeshlab>=2023.12.post2",
    "python-telegram-bot>=21.10",
    "pyvista>=0.44.2",
    "rembg>=2.0.61",
    "setuptools>=75.8.0",
    "torch>=2.5.1",
    "torchvision>=0.20.1",
    "tqdm>=4.67.1",
    "transformers>=4.48.1",
    "trimesh>=4.6.0",
    "uvicorn>=0.34.0",
    "xatlas>=0.0.9",
]
